@@ -14,6 +14,7 @@ use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
 
 /**
  * Ergast import command.
@@ -33,6 +34,7 @@ class ErgastImportCommand extends ContainerAwareCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $io = new SymfonyStyle($input, $output);
         $year = $input->getArgument('year');
 
         if ('current' == $year) {
@@ -43,16 +45,15 @@ class ErgastImportCommand extends ContainerAwareCommand
             $years = [(int) $year];
         }
 
-        /* @var $loader Importer */
-        $importer = $this->getContainer()->get('ergast_importer');
+        $importer = $this->getContainer()->get('app.ergast.importer');
 
         foreach ($years as $year) {
-            $output->writeln(sprintf('Importing the %d season...', $year));
+            $io->title(sprintf('%d Formula One season', $year));
             $s = microtime(true);
             $importer->import($year);
             $e = microtime(true);
-            $output->writeln("Memory usage after: ".(memory_get_usage() / 1024)." KB");
-            $output->writeln("in ".($e - $s)." secondes");
+            $io->comment("Memory usage after: ".(memory_get_usage() / 1024)." KB");
+            $io->comment("in ". intval($e - $s)." secondes");
         }
     }
 }
