@@ -10,7 +10,6 @@
 namespace AppBundle\Repository;
 
 use AppBundle\Entity\Constructor;
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityRepository;
 
 /**
@@ -31,16 +30,9 @@ class ConstructorRepository extends EntityRepository implements ConstructorRepos
     /**
      * {@inheritdoc}
      */
-    public function findBySlugs(array $slugs)
+    public function save(Constructor $constructor)
     {
-        $builder = $this->_em->createQueryBuilder();
-        $builder
-            ->select('c')
-            ->from($this->_entityName, 'c', 'c.slug')
-            ->where($builder->expr()->in('c.slug', ':slugs'))
-            ->setParameter(':slugs', $slugs)
-        ;
-
-        return new ArrayCollection($builder->getQuery()->getResult());
+        $this->_em->persist($constructor);
+        $this->_em->flush();
     }
 }
