@@ -9,11 +9,10 @@
 
 namespace AppBundle\Service;
 
+use AppBundle\Entity\Team;
 use Doctrine\Common\Collections\ArrayCollection;
 
 /**
- * The team service.
- *
  * @author Brieuc Thomas <tbrieuc@gmail.com>
  */
 interface TeamServiceInterface
@@ -26,4 +25,11 @@ interface TeamServiceInterface
      * @return ArrayCollection A collection of Team entities
      */
     public function findByDriver($driverId);
+
+    /**
+     * Saves a team.
+     *
+     * @param Team $team The team to save
+     */
+    public function save(Team $team);
 }
